Despliegue local terminado

+++ src/main/java/edu/eci/cvds/sampleprj/dao/mybatis/mappers/ClienteMapper.java
@@ -22,7 +22,7 @@ public interface ClienteMapper {
      * @param fechainicio
      * @param fechafin 
      */
-    public void agregarItemRentadoACliente(@Param("idc")int id, 
+    public void agregarItemRentadoACliente(@Param("idc")long id, 
             @Param("idit")int idit, 
             @Param("fechai")Date fechainicio,
             @Param("fechaf")Date fechafin);
